refactor(chat): extract chat_manager/deps.rs (#292)
Moves the ChatManagerDeps trait — the crate's dependency-injection
surface — into its own submodule, re-exported unchanged.

Task 3 of the chat_manager.rs file-limit split plan.
